LayoutTests/fast/dom/node-filter-interface-expected.txt

List NodeFilter's properties to detect changes made to it.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


* FILTER_ACCEPT
[Configurable: false, enumerable: true, writable: false, value: 1]
* FILTER_REJECT
[Configurable: false, enumerable: true, writable: false, value: 2]
* FILTER_SKIP
[Configurable: false, enumerable: true, writable: false, value: 3]
* SHOW_ALL
[Configurable: false, enumerable: true, writable: false, value: 4294967295]
* SHOW_ATTRIBUTE
[Configurable: false, enumerable: true, writable: false, value: 2]
* SHOW_CDATA_SECTION
[Configurable: false, enumerable: true, writable: false, value: 8]
* SHOW_COMMENT
[Configurable: false, enumerable: true, writable: false, value: 128]
* SHOW_DOCUMENT
[Configurable: false, enumerable: true, writable: false, value: 256]
* SHOW_DOCUMENT_FRAGMENT
[Configurable: false, enumerable: true, writable: false, value: 1024]
* SHOW_DOCUMENT_TYPE
[Configurable: false, enumerable: true, writable: false, value: 512]
* SHOW_ELEMENT
[Configurable: false, enumerable: true, writable: false, value: 1]
* SHOW_ENTITY
[Configurable: false, enumerable: true, writable: false, value: 32]
* SHOW_ENTITY_REFERENCE
[Configurable: false, enumerable: true, writable: false, value: 16]
* SHOW_NOTATION
[Configurable: false, enumerable: true, writable: false, value: 2048]
* SHOW_PROCESSING_INSTRUCTION
[Configurable: false, enumerable: true, writable: false, value: 64]
* SHOW_TEXT
[Configurable: false, enumerable: true, writable: false, value: 4]
* length
[Configurable: true, enumerable: false, writable: false, value: 0]
* name
[Configurable: true, enumerable: false, writable: false, value: NodeFilter]
PASS successfullyParsed is true

TEST COMPLETE
